Replace gitUtils.getRemoteRef with gitUtils.getRemote
Unfortunately, rev-parse only works when the name references a valid
commit object.  If the remote has not been fetched or there is a branch
name mismatch or any of several other conditions, it will not return a
remote ref for the named branch.

Replace it with git config, which only gets the remote name (that is all
that is required anyway).  This is simpler, but requires the branch name
as an argument and can't default to the current branch.  Rework the
calling logic to pass the branch.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>