Move project querying into getLastBuildInternal
Since getStatusBadge and getStatus do not require options.project
directly, there is no need to find it in canonicalizeOptions.  The
previous logic attempted to duplicate the logic in getStatus, but did
not handle the case that getLastBuild was called directly without
options.project or options.commit.  Fix this by moving it into
getLastBuildInternal.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>