Create and use AmbiguousProjectError
For when no single project is matched unambiguously.  Unlike many of the
other errors that can not be fixed automatically by the caller, this
error is one that the caller may want to handle specially (e.g. to
prompt the user).  Therefore, make it recognizable (with a special type)
and easy to use (with the .projects property of projects that matched).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>