Rework wait tests for nock 13
nock@13 makes a call to setImmediate and two calls to setTimeout for
every (non-aborted) request.  This breaks the (fragile) time mocking
test logic.  To fix this issue and avoid future issues, only mock time
in the module under test by injecting the fake timers using proxyquire
and wait for mocked requests to propagate as necessary.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>