test: Avoid depending on init.defaultBranch
The branch name created by `git init` may not be `master`, either due to
global or user configuration, or due to changes in the git default
warned about in git-init(1).  Specify --initial-branch to avoid relying
on the default (with fall-back for old versions of git which don't
support the --initial-branch option).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>