exec-file-out: reject for anything in stderr
The whitespace detection is costly (due to converting a possibly-large
Buffer to string, searching for Unicode spaces at either end, cloning
string as part of trim), sketchy (converting Buffer of unknown encoding
to string using UTF-8), and currently unused (since the commands we run
never write only whitespace to stderr on success).

Remove it.  We can consider how to re-add if a need arises.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>