refactor: Remove useProjectBuilds
This "internal option" was a bit of a kludge, which was confusing and
required special handling during wait/retry.  Instead, check if the
build can be used in getLastBuildInternal and avoid calling
getLastBuildForProject if not necessary.

Note: This change causes getProjectLastBuild{,Branch} to be called
immediately after getProjects if wait is requested and the build
requires waiting, rather than waiting RETRY_DELAY_MIN_MS.  Although this
is not ideal (assuming builds in getProjects are fresh), I think the
simplicity is justified.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>