Wait before first retry
Wait a bit before querying build status if build from project requires
waiting.  It seems reasonable to assume that if the build was not
finished when fetching the project, it is unlikely to be finished when
fetching the build status immediately afterwards.  This also matches the
behavior before 0757483 more closely.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>