test: disable test for IPv6 with zone identifier
The built-in URL class dropped support for zone identifiers in IPv6
address literals in Node.js 20.  Calling the URL constructor with a URL
containing a zone identifier causes ERR_INVALID_URL to be thrown.  This
is likely a result of switching to version 2.0 of the Ada URL parser in
<https://github.com/nodejs/node/pull/47339>.  The behavior aligns with
how [IPv6 address is defined in the WHATWG URL
Standard](https://url.spec.whatwg.org/#concept-ipv6), which notes that

> Support for <zone_id> is intentionally omitted.

As explained in the issue tracker:

https://www.w3.org/Bugs/Public/show_bug.cgi?id=27234#c2
https://github.com/whatwg/url/issues/392

Skip this test, since this URL format is not supported.  If it's
necessary to support SCP-like git URLs with zone identifiers, we'll need
to roll our own support.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>